Solved, however it has a dirty fix.
Rather than using a hashset to make sure that, that image cant be changed after, it should use the commented out part. (in gameController's setAIStatus)
But the commented out part misses a bit to work. The path that the imageUrl returns doesnt match the getResource path. Make these two match and the hashset wont be needed